% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sqlite.R
\docType{data}
\name{DISPLACE_DB_TABLES}
\alias{DISPLACE_DB_TABLES}
\title{Known DISPLACE output database tables}
\format{
A character vector.
}
\usage{
DISPLACE_DB_TABLES
}
\description{
The table names created by `SQLiteOutputStorage`, as of output schema
version 4. Use [displace_db_tables()] to see what a particular database
actually contains.
}
\examples{
DISPLACE_DB_TABLES
}
\keyword{datasets}
